[feat] #8 db User 모델 schema, 모델 분리, auth service수정
- User 모델 스키마, 기능 분리
  - User 스키마
    - 깃허브 login은 password가 없음
    - password의 required 옵션을 해제
- models/user 로 한 번 더 분리
- auth
  - github 로그인 시에 password 정보를 undefined로 넣어 추가
  - github 로그인 시에 unique 값으로 준 id가 겹칠 우려가 있어 _github를 추가
