+++ BE/src/services/auth.ts
@@ -43,11 +43,18 @@ const getGithubUserInfo = async (accessToken: string) => {
       Authorization: `BEARER ${accessToken}`,
     },
   });
-  const userInfo = { id: data.login, name: data.name };
+  const userInfo = {
+    id: `${data.login}_github`,
+    name: data.name,
+    password: undefined,
+  };
   return userInfo;
 };
 
-const checkUserInDB = async (userInfo: { id: string }): Promise<boolean> => {
+const checkUserInDB = async (userInfo: {
+  id: string;
+  password: string | undefined;
+}): Promise<boolean> => {
   const user = await userModel.get(userInfo);
   return !!user;
 };
